Hackathon/requirement.txt
tweepy
kafka-python
pandas
matplotlib
flask
kafka-python
opencv-python
tensorflow
numpy